create `Branch` trait
move old `Branch` struct to `BaseBranch`,
which implements `Branch`
