+++ src/net/architectures.rs
@@ -1,23 +1,26 @@
 use super::{
+    branch::branch::Branch,
     branch::branch::BranchCfg,
     branch::branch_cfg_builder::BranchCfgBuilder,
     net::{Net, OutputBias},
     train_stats::TrainingStats,
 };
+use std::marker::PhantomData;
 
 /// Number of markers per branch: dynamic
 /// Depth of branches: same for all branches
 /// Width of branch layers: same within branches, dynamic between branches
-pub struct BlockNetCfg {
+pub struct BlockNetCfg<B: Branch> {
     num_markers: Vec<usize>,
     depth: usize,
     widths: Vec<usize>,
     precision_prior_shape: f64,
     precision_prior_scale: f64,
     initial_random_range: f64,
+    branch_type: PhantomData<B>,
 }
 
-impl BlockNetCfg {
+impl<B: Branch> BlockNetCfg<B> {
     pub fn new() -> Self {
         BlockNetCfg {
             num_markers: vec![],
@@ -28,6 +31,7 @@ impl BlockNetCfg {
             precision_prior_shape: 1.,
             precision_prior_scale: 1.,
             initial_random_range: 0.05,
+            branch_type: PhantomData,
         }
     }
 
@@ -52,7 +56,7 @@ impl BlockNetCfg {
         self
     }
 
-    pub fn build_net(&self) -> Net {
+    pub fn build_net(&self) -> Net<B> {
         let mut branch_cfgs: Vec<BranchCfg> = Vec::new();
         let num_branches = self.widths.len();
         for branch_ix in 0..num_branches {
@@ -75,17 +79,19 @@ impl BlockNetCfg {
             },
             error_precision: 1.0,
             training_stats: TrainingStats::new(),
+            branch_type: PhantomData,
         }
     }
 }
 
 #[cfg(test)]
 mod tests {
+    use super::super::branch::base_branch::BaseBranch;
     use super::BlockNetCfg;
 
     #[test]
     fn test_block_net_architecture_num_params_in_branch() {
-        let mut cfg = BlockNetCfg::new().with_depth(1);
+        let mut cfg = BlockNetCfg::<BaseBranch>::new().with_depth(1);
         cfg.add_branch(3, 3);
         cfg.add_branch(3, 3);
         let net = cfg.build_net();
