+++ src/net/branch/mod.rs
@@ -1,3 +1,4 @@
+pub mod base_branch;
 pub mod branch;
 pub mod branch_builder;
 pub mod branch_cfg_builder;
