+++ src/net/branch/momenta.rs
@@ -1,4 +1,4 @@
-use super::branch::BranchLogDensityGradient;
+use super::base_branch::BranchLogDensityGradient;
 use super::step_sizes::StepSizes;
 use arrayfire::Array;
